Refactor part of the interfaces and methods
